Delete parts of the crate-level docs.
This is attempting to make the crate docs less overwhelming.

This removes the explanation for the global lock, assuming
the reader will understand the implications of fail points
being global resources.

It also removes a big explanation about the nature of the
closure passed to the macro. This information is also
in the docs for the macro itself.

It also removes a minor style suggestion from the
final bullet list of considerations.

Fixes #27.