Support customized logic for FAILPOINTS tasks
Previously, the only way to provide customized (as a Rust function)
failpoint behavior is to use `cfg_callback` API. This change extends
the customization to the FAILPOINTS environment variable. Application
can provide a function to turn a `call(arg)` task in FAILPOINTS to
a "callback". Unlike `cfg_callback`, this integrates with the rest
of FAILPOINTS syntax, for example:

    3*call(foo)->2*return->panic

`cfg_callback` does not seem to be able to express `return`.

The use-case that motivates this change is to test race conditions
in multiple processes. I'd like to use 2 primitives:

    wait("foo")     // wait for "foo".
    unblock("foo")  // unblock a current or future wait of "foo".

instead of "sleep" to create interesting racy cases deterministically.

Signed-off-by: Jun Wu <quark@lihdd.net>